requirements.txt
ipyKernel
openai>=1.0.0
python-dotenv
yfinance
pandas
matplotlib